requirements.txt
Flask>=2.3.0
Flask-Cors>=4.0.0
gunicorn>=21.2.0
soundfile>=0.12.1
numpy>=1.24.0
openai-whisper
torch

